+++ src/Igor/src/iNodeParticleSystem.h
@@ -152,15 +152,15 @@ namespace Igor
         \param columns column count (<= 1)
         \param rows row count (<= 1)
         */
-        void setFirstTextureTiling(uint32 columns, uint32 rows);
+        void setFirstTextureTiling(uint8 columns, uint8 rows);
 
         /*! \returns the horizontal tiling resolution of the first texture layer
         */
-        uint32 getFirstTextureColumns() const;
+        uint8 getFirstTextureColumns() const;
 
         /*! \returns the vertical tiling resolution of the first texture layer
         */
-        uint32 getFirstTextureRows() const;
+        uint8 getFirstTextureRows() const;
 
         /*! sets rainbow gradient for particles color
